requirements.txt
streamlit~=1.30.0
tensorflow
numpy~=1.26.4
pandas~=2.2.2
scikit-learn
fastapi~=0.114.0
uvicorn~=0.30.6
python-multipart
pillow~=10.3.0
pydantic~=1.10.12